requirements.txt
torch==1.13.1
PyYAML==6.0
opencv-python==4.7.0.68
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
seaborn==0.12.2
Pillow==9.4.0
scipy==1.7.3
torchvision==0.14.1
tqdm==4.64.1
thop==0.1.1.post2209072238
requests==2.28.1
celery==5.2.7
Flask==2.2.2
imagezmq==1.1.1
pyserial==3.5
imutils==0.5.4
pyrealsense2==2.53.1.4623